review/git-review: rewrite commit-msg hook in Go
We are surely going to need to change this in the future.
Make the commit-hook script as trivial as possible, just
delegating to the git-review command, so that we have
some hope of making changes when we need to.

This should also be a good pattern for the gofmt hook.

Change-Id: Ia9cab1320cf50c12c1766a91f7164696b68d8922
Reviewed-on: https://go-review.googlesource.com/1584
Reviewed-by: Andrew Gerrand <adg@golang.org>