git-codereview: make Branch.Submitted pickier
Insist that the search term show up in the output.
This prevents unrelated junk output
from generating false positives.

Treat the search term as a fixed string
rather than a regular expression,
since that's what it is.

Limit the number of matches to 1.
git log is fast, but why waste cycles?

Updates golang/go#9468

Change-Id: I3ff1ab48785234481543c524a99270b9ecc0ed8e
Reviewed-on: https://go-review.googlesource.com/2155
Reviewed-by: Chris Manghane <cmang@golang.org>
Reviewed-by: Andrew Gerrand <adg@golang.org>