git-codereview: add mail -topic support
This adds a "-topic" flag to the mail subcommand that sets the Gerrit
topic for the pushed changes.  Gerrit topics make it easier to group
related changes such as in a multi-change review.

Note that, like reviewers, the topic is sticky, so it is only
necessary to use mail -topic once for a given change.

Change-Id: I30cf9a88092cc9270b4643d432dec7e5f967b922
Reviewed-on: https://go-review.googlesource.com/5301
Reviewed-by: Russ Cox <rsc@golang.org>