git-codereview: Properly detect Rietveld-style fix messages
Detection of Rietveld-style fix messages is already in place as a warning,
but was broken when we added feature to rewrite issue references in commit edbdf1a7 .
To fix this, we automatically reformat the fix message into the GitHub style.

We also forget to write the change to the fix message back,
if we do not also need to write Change-Id and origin branch.
This change also fixes this.

Fixes golang/go#11472

Change-Id: Ie77358867e38cf976a0688b6e2f80525dae3891e
Reviewed-on: https://go-review.googlesource.com/15754
Reviewed-by: Andrew Gerrand <adg@golang.org>