git-codereview: fix commitmsg hook when verbose
If the commit message is verbose it shows the diff being committed. This
can be shown by either `git commit --verbose` or setting the following
in your git config:

	[commit]
		verbose = true

The diff is uncomented and shown after a "everything below will be
removed" comment section. Identify it and do remove everything under it,
as otherwise the diff would get mixed up with the commit message.

Fixes golang/go#16376.

Change-Id: Ia69aeaf36a1c8471da423a142fb233168d455a5d
Reviewed-on: https://go-review.googlesource.com/25342
Reviewed-by: Josh Bleecher Snyder <josharian@gmail.com>