review/git-codereview: use more verbosity to signal verbose pulls
The -v flag in review is already used to signal more verbose output.
With one -v, git sync shows the git commands. After this change,
two -vs (or -v=2) causes git sync to pass -v to git pull as well.

Also mention that -v is a counter not a boolean in the documentation.

Change-Id: I584068f05c2298be6ae2e6ae17a0b817dc6a092f
Reviewed-on: https://go-review.googlesource.com/88875
Reviewed-by: Andrew Gerrand <adg@golang.org>