git-codereview: use --no-ff for sync-branch
When working with dev branches, it is possible that a branch
needed to be sync'd can be fast-forworded. For example, there is
no new commit since the last sync. Or this could happen with
multi-release-cycle branches, where we reverse-sync'd to master
for one release cycle, later try to sync master to branch for the
next cycle. Fast-forwording is generally not what we want (and it
confuses the git-codereview tool when trying to craft the CL
description). Pass --no-ff to ensure we create a merge commit.

Change-Id: Ic204882d0304ec538c2e219d34e370ba98db9374
Reviewed-on: https://go-review.googlesource.com/c/review/+/319069
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@google.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>