git-codereview: frame sync as git codereview sync
The codereview binary referred to the sync subcommand as merely "git
sync" in a number of places in user surfaces.  This was wrong and
misleading, so the code now refers to it with the codereview parent
subcommand in the name (i.e., a "git codereview sync").

Change-Id: Icf3f96a8c76b57319bc3101b3e9fd0cbd8cecea9
Reviewed-on: https://go-review.googlesource.com/c/review/+/685855
Auto-Submit: Sean Liao <sean@liao.dev>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Carlos Amedee <carlos@golang.org>
Reviewed-by: Sean Liao <sean@liao.dev>
Reviewed-by: David Chase <drchase@google.com>